+++ src/main/java/com/nhnacademy/bookstoreuserapi/orderPointProcess/service/OrderPointProcessService.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.bookstoreuserapi.orderPointProcess.service;
 
-import com.nhnacademy.bookstoreuserapi.orderPointProcess.domain.request.OrderPointMinusProcessRequest;
-import com.nhnacademy.bookstoreuserapi.orderPointProcess.domain.request.OrderPointPlusProcessRequest;
+import com.nhnacademy.bookstoreuserapi.orderPointProcess.dto.request.OrderPointMinusProcessRequest;
+import com.nhnacademy.bookstoreuserapi.orderPointProcess.dto.request.OrderPointPlusProcessRequest;
 
 public interface OrderPointProcessService {
